+++ Punto.h
@@ -17,17 +17,17 @@ class Punto
 {
 private:
     int indice;
-    int x;
-    int y;
-    int z;
+    double x;
+    double y;
+    double z;
 //    vector <Objeto *> esta_en_objetos;
 //    vector <Linea *> esta_en_aristas;
 //    vector <Plano *> esta_en_caras;
 public:
-    Punto(int indice,int x, int y, int z);
-    int get_x();
-    int get_y();
-    int get_z();
+    Punto(int indice,double x, double y, double z);
+    double get_x();
+    double get_y();
+    double get_z();
 /*     Objeto * get_esta_en_objetos(int indice);
     vector<Objeto *> get_esta_en_objetos();
     Linea* get_esta_en_aristas(int indice);
